ajustando ids da classe acount

+++ src/MudblazorAuth.Domain/Entities/Account.cs
@@ -7,11 +7,11 @@ namespace MudblazorAuth.Domain.Entities
 	{
 		[Key]
 		[DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public int Id { get; set; }
+		public long Id { get; set; }
 		public string Username { get; set; } = string.Empty;
 		public string Password { get; set; } = string.Empty;
 		[ForeignKey("Profile")]
-		public int IdProfile { get; set; }
+		public long IdProfile { get; set; }
 		public Profile Profile { get; set; } = default!;
 	}
 }
